temp

+++ src/GraphGAN/discriminator.py
@@ -14,9 +14,9 @@ class Discriminator(object):
         self.node_neighbor_id = 0
         self.reward = 0
 
-        self.node_embedding = torch.index_select(self.embedding_matrix, 0, self.node_id)
-        self.node_neighbor_embedding = torch.index_select(self.embedding_matrix, 0, self.node_neighbor_id)
-        self.bias = torch.index_select(self.bias_vector, 0, self.node_neighbor_id)
+        self.node_embedding = torch.index_select(self.embedding_matrix, 0, self.node_id.long())
+        self.node_neighbor_embedding = torch.index_select(self.embedding_matrix, 0, self.node_neighbor_id.long())
+        self.bias = torch.index_select(self.bias_vector, 0, self.node_neighbor_id.long())
         self.score = self.node_embedding*self.node_neighbor_embedding.sum(0) + self.bias
